Invoices L1,L2,L3

+++ src/app/pages/logout/logout.component.ts
@@ -3,6 +3,7 @@ import { Router } from "@angular/router";
 import { NbDialogService } from "@nebular/theme";
 import { AuthService } from "../../services/auth.service";
 import { ShowcaseDialogComponent } from "../modal-overlays/dialog/showcase-dialog/showcase-dialog.component";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "ngx-logout",
@@ -17,19 +18,25 @@ export class LogoutComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    this.open();
+    this.openSwalFire();
   }
-  open() {
-    this.dialogService
-      .open(ShowcaseDialogComponent, {
-        context: {
-          title: "Logout",
-          body: "Are you sure you want to logout?",
-        },
-      })
-      .onClose.subscribe(() => {
-        sessionStorage.clear();
-        this.router.navigate(["/dashboard"]);
+  openSwalFire() {
+      Swal.fire({
+        title: "Are you sure you want to Logout?",
+        text: "You will be Logged out of this Application!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, Logout !",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          sessionStorage.clear();
+          this.router.navigate(["/login"]);
+        } else {
+          console.log("You Clicked No");
+          this.router.navigate(["/pages/dashboard"]);  
+        }
       });
   }
 }
